Match Serial prototype

+++ Marlin/src/HAL/HAL_AVR/MarlinSerial.cpp
@@ -50,7 +50,7 @@
   template<typename Cfg> uint8_t  MarlinSerial<Cfg>::rx_dropped_bytes = 0;
   template<typename Cfg> uint8_t  MarlinSerial<Cfg>::rx_buffer_overruns = 0;
   template<typename Cfg> uint8_t  MarlinSerial<Cfg>::rx_framing_errors = 0;
-  template<typename Cfg> uint32_t MarlinSerial<Cfg>::current_baudrate = BAUDRATE;
+  template<typename Cfg> long MarlinSerial<Cfg>::current_baudrate = BAUDRATE;
   template<typename Cfg> typename MarlinSerial<Cfg>::ring_buffer_pos_t MarlinSerial<Cfg>::rx_max_enqueued = 0;
 
   // A SW memory barrier, to ensure GCC does not overoptimize loops
@@ -325,12 +325,12 @@
 
   // Public Methods
   template<typename Cfg>
-  uint32_t MarlinSerial<Cfg>::baudrate() {
+  long MarlinSerial<Cfg>::baudrate() {
     return current_baudrate;
   }
 
   template<typename Cfg>
-  void MarlinSerial<Cfg>::begin(const uint32_t baud) {
+  void MarlinSerial<Cfg>::begin(const long baud) {
     uint16_t baud_setting;
     bool useU2X = true;
     current_baudrate = baud;
